Moved filter_fields to more appropriate location
